third_party/blink/web_tests/external/wpt/html/browsers/history/the-location-interface/assign_after_load-expected.txt

This is a testharness.js-based test.
[FAIL] Assignment to location after document is completely loaded
  assert_equals: expected 3 but got 2
Harness: the test ran to completion.
